stack/backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn>=0.35.0
wsproto>=1.2.0
httptools>=0.6.0
uvloop>=0.19.0

# Redis
redis>=5.0.0

# MCP scrubbing (controls: httpx, pydantic, python-dotenv, python-multipart)
fastmcp==2.14.5

# File type detection
python-magic>=0.4.27
